添加store下modules文件夹结构
