+++ src/store/getters.js
@@ -1,0 +1,6 @@
+const getters = {
+  currentUser: state => state.user.currentUser,
+  isLogin: state => state.user.isLogin
+}
+
+export default getters
