--- src/vuex/getters.js
@@ -1,2 +1,0 @@
-export const currentUser = state => state.currentUser
-export const isLogin = state => state.isLogin
